requirements.txt
dnspython
tqdm
colorama
